Create `borrow` module.
