+++ src/lib.rs
@@ -14,6 +14,7 @@ extern crate lazy_static;
 pub mod vm;
 pub mod value;
 pub mod object;
+pub mod borrow;
 pub mod thread;
 pub mod meta;
 pub mod prelude;
