+++ src/object/class/mod.rs
@@ -8,7 +8,8 @@ use neon_runtime;
 use neon_runtime::raw;
 use vm::{JsResult, VmResult, Throw, Context, VmGuard, Callback};
 use vm::internal::Isolate;
-use value::{Value, JsFunction, JsValue, Handle, Managed, Borrow, BorrowMut, Ref, RefMut, LoanError, build};
+use borrow::{Borrow, BorrowMut, Ref, RefMut, LoanError};
+use value::{Value, JsFunction, JsValue, Handle, Managed, build};
 use value::internal::ValueInternal;
 use object::{Object, This};
 use self::internal::{ClassMetadata, MethodCallback, ConstructorCallCallback, AllocateCallback, ConstructCallback};
