+++ src/prelude.rs
@@ -1,5 +1,6 @@
 //! The Neon "prelude," a re-exported collection of the most commonly-used Neon APIs.
 
-pub use value::{Handle, JsBuffer, JsArrayBuffer, BinaryData, JsError, ErrorKind, Value, JsValue, JsUndefined, JsNull, JsBoolean, JsString, ToJsString, JsNumber, JsObject, JsArray, JsFunction, Borrow, BorrowMut};
+pub use value::{Handle, JsBuffer, JsArrayBuffer, BinaryData, JsError, ErrorKind, Value, JsValue, JsUndefined, JsNull, JsBoolean, JsString, ToJsString, JsNumber, JsObject, JsArray, JsFunction};
 pub use object::{Object, Class};
+pub use borrow::{Borrow, BorrowMut};
 pub use vm::{VmResult, JsResult, JsResultExt, CallKind, Context, ModuleContext, ExecuteContext, ComputeContext, CallContext, FunctionContext, MethodContext};
