+++ src/value/mod.rs
@@ -8,11 +8,10 @@ use std;
 use std::fmt;
 use std::os::raw::c_void;
 use std::marker::PhantomData;
-use std::ops::{Deref, DerefMut, Drop};
 use neon_runtime;
 use neon_runtime::raw;
-use vm::{Context, VmGuard, FunctionContext, Callback, VmResult, Throw, JsResult, JsResultExt};
-use vm::internal::{Isolate, Pointer};
+use vm::{Context, FunctionContext, Callback, VmResult, Throw, JsResult, JsResultExt};
+use vm::internal::Isolate;
 use object::{Object, This};
 use self::internal::{ValueInternal, SuperType, FunctionCallback};
 
@@ -621,136 +620,5 @@ impl<T: Object> ValueInternal for JsFunction<T> {
 
     fn is_typeof<Other: Value>(other: Other) -> bool {
         unsafe { neon_runtime::tag::is_function(other.to_raw()) }
-    }
-}
-
-/// A trait for JS values whose internal contents can be borrowed immutably by Rust while the JS VM is locked.
-pub trait Borrow: Sized {
-
-    /// The type of the value's internal contents.
-    type Target: Pointer;
-
-    /// Borrow the contents of this value immutably.
-    /// 
-    /// If there is already an outstanding mutable loan for this value, this method panics.
-    fn borrow<'a>(self, guard: &'a VmGuard<'a>) -> Ref<'a, Self::Target> {
-        match self.try_borrow(guard) {
-            Ok(r) => r,
-            Err(e) => panic!("{}", e)
-        }
-    }
-
-    /// Borrow the contents of this value immutably.
-    /// 
-    /// If there is already an outstanding mutable loan for this value, this method fails with a `LoanError`.
-    fn try_borrow<'a>(self, guard: &'a VmGuard<'a>) -> Result<Ref<'a, Self::Target>, LoanError>;
-
-}
-
-/// A trait for JS values whose internal contents can be borrowed mutably by Rust while the JS VM is locked.
-pub trait BorrowMut: Borrow {
-
-    /// Borrow the contents of this value mutably.
-    /// 
-    /// If there is already an outstanding loan for this value, this method panics.
-    fn borrow_mut<'a>(self, guard: &'a VmGuard<'a>) -> RefMut<'a, Self::Target> {
-        match self.try_borrow_mut(guard) {
-            Ok(r) => r,
-            Err(e) => panic!("{}", e)
-        }
-    }
-
-    /// Borrow the contents of this value mutably.
-    /// 
-    /// If there is already an outstanding loan for this value, this method panics.
-    fn try_borrow_mut<'a>(self, guard: &'a VmGuard<'a>) -> Result<RefMut<'a, Self::Target>, LoanError>;
-
-}
-
-/// An error produced by a failed loan in the `Borrow` or `BorrowMut` traits.
-pub enum LoanError {
-
-    /// Indicates that there is already an outstanding mutable loan for the object at this address.
-    Mutating(*const c_void),
-
-    /// Indicates that there is already an outstanding immutable loan for the object at this address.
-    Frozen(*const c_void)
-
-}
-
-impl fmt::Display for LoanError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match *self {
-            LoanError::Mutating(p) => {
-                write!(f, "outstanding mutable loan exists for object at {:?}", p)
-            }
-            LoanError::Frozen(p) => {
-                write!(f, "object at {:?} is frozen", p)
-            }
-        }
-    }
-}
-
-/// An immutable reference to the contents of a borrowed JS value.
-pub struct Ref<'a, T: Pointer> {
-    pointer: T,
-    guard: &'a VmGuard<'a>
-}
-
-impl<'a, T: Pointer> Ref<'a, T> {
-    pub(crate) unsafe fn new(guard: &'a VmGuard<'a>, pointer: T) -> Result<Self, LoanError> {
-        let mut ledger = guard.ledger.borrow_mut();
-        ledger.try_borrow(pointer.as_ptr())?;
-        Ok(Ref { pointer, guard })
-    }
-}
-
-impl<'a, T: Pointer> Drop for Ref<'a, T> {
-    fn drop(&mut self) {
-        let mut ledger = self.guard.ledger.borrow_mut();
-        ledger.settle(unsafe { self.pointer.as_ptr() });
-    }
-}
-
-impl<'a, T: Pointer> Deref for Ref<'a, T> {
-    type Target = T;
-
-    fn deref(&self) -> &Self::Target {
-        &self.pointer
-    }
-}
-
-/// A mutable reference to the contents of a borrowed JS value.
-pub struct RefMut<'a, T: Pointer> {
-    pointer: T,
-    guard: &'a VmGuard<'a>
-}
-
-impl<'a, T: Pointer> RefMut<'a, T> {
-    pub(crate) unsafe fn new(guard: &'a VmGuard<'a>, mut pointer: T) -> Result<Self, LoanError> {
-        let mut ledger = guard.ledger.borrow_mut();
-        ledger.try_borrow_mut(pointer.as_mut())?;
-        Ok(RefMut { pointer, guard })
-    }
-}
-
-impl<'a, T: Pointer> Drop for RefMut<'a, T> {
-    fn drop(&mut self) {
-        let mut ledger = self.guard.ledger.borrow_mut();
-        ledger.settle_mut(unsafe { self.pointer.as_mut() });
-    }
-}
-
-impl<'a, T: Pointer> Deref for RefMut<'a, T> {
-    type Target = T;
-
-    fn deref(&self) -> &Self::Target {
-        &self.pointer
-    }
-}
-
-impl<'a, T: Pointer> DerefMut for RefMut<'a, T> {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.pointer
     }
 }
